requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
google-generativeai
python-dotenv
pydantic